# app/Http/Requests/StoreTransactionRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class StoreTransactionRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     *
     * @return bool
     */
    public function authorize()
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array
     */
    public function rules()
    {
        return [
            'user_id' => [
                'required',
            ],
            'date_transaction' => [
                'required',
            ],
            'description' => [
                'required',
            ],
            'fl_operation' => [
                'required',
            ],
            'fl_reversal' => [
                'required',
            ],
            'value' => [
                'required',
            ],
        ];
    }
}